+++ smoke_tests/commonjs/commonjs.js
@@ -8,8 +8,10 @@ const {
   config,
   getJson,
   getHtml,
+  getMd,
   getJsonBySearchId,
   getHtmlBySearchId,
+  getMdBySearchId,
   getAccount,
   getLocations,
 } = require("serpapi");
@@ -90,6 +92,12 @@ const run = async () => {
   }
 
   {
+    console.log("getMd");
+    const markdown = await getMd(Object.assign({ engine: "google" }, params));
+    if (!markdown.startsWith("---")) throw new Error("Incorrect Markdown");
+  }
+
+  {
     console.log("getJsonBySearchId");
     config.api_key = apiKey;
     const json = await getJsonBySearchId(searchId);
@@ -109,6 +117,13 @@ const run = async () => {
     getHtmlBySearchId(searchId, {}, (html) => {
       if (html.length < 1000) throw new Error("Incorrect HTML");
     });
+  }
+
+  {
+    console.log("getMdBySearchId");
+    config.api_key = apiKey;
+    const markdown = await getMdBySearchId(searchId);
+    if (!markdown.startsWith("---")) throw new Error("Incorrect Markdown");
   }
 
   {